drm/vc4: Add DSI1 driver
The DSI0 and DSI1 blocks on the 2835 are different but very similar
hardware blocks.  Some registers move around, and the featureset is
slightly different, but they're clearly related.  This doesn't enable
DSI0 (since I don't have a connector for it on the Pi), but some of
the infrastructure is present.

Signed-off-by: Eric Anholt <eric@anholt.net>