mmc: bcm2835-sdhost: Add new driver for the internal SD controller.
The 2835 has two SD controllers: The Arasan SDHCI controller that we
currently use, and a custom SD controller.  The custom one runs faster
at 16.7MB/sec reads on PIO (according to hdparm -t on a Pi 2B) instead
of 10.15MB/sec, and up to 18.45MB/sec with DMA enabled.

The code was originally written by Phil Elwell in the downstream
Rasbperry Pi tree, and I did a major cleanup on it (+319, -721 lines
out of the original 2055) for inclusion.

Signed-off-by: Eric Anholt <eric@anholt.net>