bcm2708: Move most of vc4-kms-v3d-overlay to bcm2708_common.
The overlay's really just telling the firmware that the driver will be
in use, and the only component that the firmware reads is the
interrupts.  So, we can leave the interrupts in the overlay and move
the rest to the core include.

This will reduce bcm2708's diff to bcm2835, and let us fix the HDMI
HPD GPIO active state per board.

Signed-off-by: Eric Anholt <eric@anholt.net>