ARM: bcm2385: Don't force pl011 onto pins 14/15.
The firmware hands off to Linux with pins 14/15 configured to a uart,
but on the Raspberry Pi 3 the miniuart is used instead of pl011.  When
we reconfigured pinmux to put pl011 on these pins, we ended up with
the pl011 talking to both bluetooth and serial console, so neither
worked.

Signed-off-by: Eric Anholt <eric@anholt.net>