drm/vc4: Fix overflow mem unreferencing when the binner runs dry.
Overflow memory handling is tricky: While it's still referenced by the
BPO registers, we want to keep it from being freed.  When we are
putting a new set of overflow memory in the registers, we need to
assign the old one to the last rendering job using it.

We were looking at "what's currently running in the binner", but since
the bin submission split, we may end up with the binner completing and
having no new job.  However, a previously-binned job may still be
using the memory while it's in the queue for rendering, and we would
be freeing the memory while the device was still using it.  So, if we
don't find a bin job at all, look at the highest-seqno (last) render
job to attach our overflow to.

Signed-off-by: Eric Anholt <eric@anholt.net>
Fixes: ca26d28bbaa3 ("drm/vc4: improve throughput by pipelining binning and rendering jobs")