RFC: name the Raspberry Pi GPIO lines
This is a scratch idea on how to properly name the GPIO lines
on the various Raspberry Pi models.

The idea is to give useful names to GPIO lines that an
implementer will be using from userspace, e.g. for maker type
projects.

See:
commit 521a2ad6f862
("gpio: add userspace ABI for GPIO line information")
commit fd9c55315db9
("gpio: of: make it possible to name GPIO lines")

And watch how nice I think it looks on lsgpio (I do not have
a raspberry pi).

I need help in getting the right names for the lines and
matching the device tree to the right schematic so I don't
get this wrong.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Eric Anholt <eric@anholt.net>