ARM: bcm2835: Rename Raspberry Pi GPIO lines on the expansion header.
Doing a google image search for "Raspberry Pi GPIO", the top 5
different images used the pin number starting from 1 on the expansion
header then the SOC's GPIO number as their most visible labelings.
The "_GENx" label was only present on 1 out of the 5 images, and has
disappeared from more recent schematics.

So, label the nodes P<number> if they're on the header, and
GPIO<number> if they were previously GPIO_GENx.

In the process I found that:
- CAM_GPIO* are moved on the A+/B+.
- SD_CARD_DET is gone on the A+/B=/
- STATUS_LED moved on the A+/B+/

Signed-off-by: Eric Anholt <eric@anholt.net>