drm/vc4: Add threading support
Add support for threaded fragment shaders. The validation code checks
for thread switch signals and ensures that the registers of the other
thread are not touched. Also it tracks the live clamps so they cannot
be overwritten by the other shader.

It checks that the threading and branching instructions do not
interfere.

Signed-off-by: Jonas Pfeil <pfeiljonas@gmx.de>