dt-bindings: new binding for Himax HX8357D display panels
This adds a new binding for Himax HX8357D display panels. It includes
a compatible string for one display (more can be added in the future).

The YX350HV15 panel[1] is found in the Adafruit PiTFT 3.5" Touch
Screen for Raspberry Pi.

[1] https://learn.adafruit.com/adafruit-pitft-3-dot-5-touch-screen-for-raspberry-pi/downloads

This binding is closely modeled after the ili9341 binding, for a
similar product from adafruit.  The primary difference is that the
hx8357d doesn't have a reset line that I can find in the schematics.

v2: Document the "reg" property (requested by Rob), fix commit message
        typo (Noralf)

Signed-off-by: Eric Anholt <eric@anholt.net>
Link: https://patchwork.freedesktop.org/patch/msgid/20181024184313.2967-2-eric@anholt.net
Reviewed-by: Rob Herring <robh@kernel.org>
Acked-by: Noralf Trønnes <noralf@tronnes.org>